refactor(python code related to basefold, mle and field.py, utils.py): remove mle.py
remove mle.py, fix problems caused by this operation
